fix(tcl): Enable expression index tests by fixing persistence (#4790)
fix(tcl): enable expression index tests by fixing persistence

The indexexpr1.test file was silently failing because:

1. The TCL test harness was skipping tests with expression indexes,
   marking them as "not supported" in uses_sqlite_internals()

2. When expression indexes were created, the SQL dump persistence
   code would panic on expect_column_name() for expression indexes

3. The binary catalog persistence also panicked on expression indexes

This commit:
- Removes the expression index skip logic from tester_vibesql.tcl
- Fixes SQL dump persistence (save.rs) to properly serialize
  expression indexes using to_sql()
- Fixes binary catalog persistence with a new v6 format that stores
  index column type (0=column, 1=expression) followed by content
- Updates read_catalog_v to handle both old (v1-5) and new (v6+) formats
- Adds vibesql-parser as a dependency for parsing expression SQL on load
- Fixes an unrelated clippy warning about unnecessary f64 cast

With this fix:
- indexexpr1-100 setup test now runs and creates the expression index
- 20 tests now pass (up from 0 appearing in output)
- The remaining 86 failures are due to incomplete expression index
  support in the query executor (separate issues)

+++ crates/vibesql-ast/src/pretty_print.rs
@@ -144,7 +144,7 @@ impl ToSql for SqlValue {
                 }
             }
             SqlValue::Real(r) => {
-                let f64_val = *r as f64;
+                let f64_val = *r;
                 if f64_val.is_nan() || f64_val.is_infinite() {
                     "NULL".to_string()
                 } else if f64_val.fract() == 0.0 {
